requirements.txt
astroid
autopep8
Django
djangorestframework
djangorestframework-filters
django-rest-auth
django-social-auth
isort
lazy-object-proxy
Markdown
mccabe
Pillow
psycopg2
psycopg2-binary
pycodestyle
pylint
pytz
six
typed-ast
wrapt
social-auth-app-django
social-auth-core
oauth2
social-auth-app-django
social-auth-core
vk
fabric3
django-filters
django-compressor
django-filter==2.0.0
